+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Exclude "sun.*" objects from configuration analysis (#170, #180)
 * Fix junit5 extensions support for @TestInstance(TestInstance.Lifecycle.PER_CLASS)
   (class instance injections now processed in beforeEach hook instead of instancePostProcessor)
+* Add error message when junit5 extensions incorrectly registered with non-static fields (to avoid confusion)
 
 Lambda modules reporting:
 - Root lambda module class will be shown in the diagnostic report (in a list of root modules)